+++ src/bzflag/CacheMenu.cxx
@@ -121,7 +121,7 @@ CacheMenu::CacheMenu()
   listHUD.push_back(failedMessage);
 
 
-  initNavigation(listHUD, 1, listHUD.size() - 2);
+  initNavigation(listHUD, 1, (int)listHUD.size() - 2);
 
   return;
 }
@@ -204,7 +204,7 @@ void CacheMenu::resize(int _width, int _height)
   x = 0.5f * (float)_width;
   y -= 0.6f * titleHeight;
   const float h = fm.getStrHeight(MainMenu::getFontFace(), fontSize, " ");
-  const int count = listHUD.size();
+  const int count = (const int)listHUD.size();
   int i;
   for (i = 1; i < count; i++) {
     listHUD[i]->setFontSize(fontSize);
